Revert "copy the point to the named curve before calling `EC_POINT_is_on_curve` due to how OpenSSL 1.0.2 behaves"
This reverts commit 716e92b99630a7845cb67755460ea9e57d5f3357.

If the private key has a name in the `EC PARAMETERS` section of the PEM file, then the reverted code is not necessary.
And we want to only support that case where not the values of the parametres but the name of the curve is stored within
the key file and the certifcate, since the browsers only support certificate files that specify the curves by their names.