check handshake messages that may trigger key changes are at the end of the record
draft-19 section 5.1: Handshake messages MUST NOT span key changes. Because the ClientHello, EndOfEarlyData, ServerHello, Finished, and KeyUpdate messages can arrive immediately prior to a key change, upon receiving these messages a receiver MUST verify that the end of these messages aligns with a record boundary; if not, then it MUST terminate the connection with an “unexpected_message” alert.