Alternate fix proposal
Analyzing the code, it seems that we had two untested conditions. First, "validation recursive" expects that the `byte_max` value is > 0; it may read into unallocated memory if called with length = 0. Second, validation_recursive may return an error, which was not immediately tested. Adding those two tests *should* fix the fuzzing bug, but that needs validation.